Changed library layout
Split 'calculus.rs' into a folder, and the trait 'Calculus' into 'Intergration', 'Differentiation' and 'TermTrait', which are used to describe a type's ability to do intergration, differentiation, and acting as a term respectively-these three behaviours are combined into a new trait 'Calculus', which is automatically implemented for any type that implememnts all three types.

The three traits are now automatically implemented for a Vec of any type that implements it's respective trait, which is an improvement over before, when it was only for a Vec of Terms.